pyproject.toml
# /apps/backend-fastapi/pyproject.toml

[tool.poetry]
name = "PyNote"
version = "0.0.1"
description = "Inbound Knowledge Manager"
authors = ["Jaehak Lee <leejaehak87@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python=">=3.10, <3.13"
pyside6 = "^6.9.1"
requests = "^2.32.4"
beautifulsoup4 = "^4.13.4"
feedparser = "^6.0.11"
python-dotenv = "^1.1.1"
pyyaml = "^6.0.2"
google-api-python-client = "^2.179.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
tzdata = "^2025.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"